move hashkey code into board.c/h
