+++ src/makemove.c
@@ -4,7 +4,6 @@
 #include "bitboards.h"
 #include "board.h"
 #include "psqt.h"
-#include "hashkeys.h"
 #include "move.h"
 #include "validate.h"
